Start cleanup of units.
