+++ lib/alchemist/si_units.rb
@@ -1,0 +1,13 @@
+module Alchemist
+  def self.si_units
+    %w[
+      m meter metre meters metres liter litre litres liters l L
+      farad farads F coulombs C gray grays Gy siemen siemens S
+      mhos mho ohm ohms volt volts V joule joules J newton
+      newtons N lux lx henry henrys H b B bits bytes bit byte
+      lumen lumens lm candela candelas cd tesla teslas T gauss
+      Gs G gram gramme grams grammes g watt watts W pascal
+      pascals Pa becquerel becquerels Bq curie curies Ci
+    ]
+  end
+end
